Rename scripts and cleanup package.json
