+++ backend/src/scripts/fetchMovies.ts
