+++ backend/src/scripts/fetchPersons.ts
